@@ -1,69 +1,253 @@
 {
   "name": "superagent",
-  "version": "4.1.0",
-  "description": "elegant & feature rich browser / node HTTP with a fluent API",
-  "scripts": {
-    "prepare": "make all",
-    "test": "make test",
-    "test-http2": "make test-node-http2"
+  "description": "Small progressive client-side HTTP request library, and Node.js module with the same API, sporting many high-level HTTP client features",
+  "version": "5.0.0",
+  "author": "TJ Holowaychuk <tj@vision-media.ca> (http://tjholowaychuk.com)",
+  "ava": {
+    "failFast": true,
+    "verbose": true
   },
-  "keywords": [
-    "http",
-    "ajax",
-    "request",
-    "agent"
-  ],
-  "license": "MIT",
-  "author": "TJ Holowaychuk <tj@vision-media.ca>",
+  "browser": {
+    "./lib/node/request.js": "./lib/client.js",
+    "./test/support/server.js": "./test/support/blank.js"
+  },
+  "bugs": {
+    "url": "https://github.com/visionmedia/superagent/issues",
+    "email": "tj@vision-media.ca"
+  },
+  "commitlint": {
+    "extends": [
+      "@commitlint/config-conventional"
+    ]
+  },
   "contributors": [
+    "TJ Holowaychuk <tj@vision-media.ca> (http://tjholowaychuk.com)",
     "Kornel Lesiński <kornel@geekhood.net>",
     "Peter Lyons <pete@peterlyons.com>",
-    "Hunter Loftis <hunter@hunterloftis.com>"
+    "Hunter Loftis <hunter@hunterloftis.com>",
+    "Nick Baugh <niftylettuce@gmail.com>"
   ],
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/visionmedia/superagent.git"
-  },
   "dependencies": {
-    "component-emitter": "^1.2.0",
+    "component-emitter": "^1.2.1",
     "cookiejar": "^2.1.2",
-    "debug": "^4.1.0",
-    "form-data": "^2.3.3",
-    "formidable": "^1.2.0",
-    "methods": "^1.1.1",
-    "mime": "^2.4.0",
-    "qs": "^6.6.0",
-    "readable-stream": "^3.0.6"
+    "debug": "^4.0.1",
+    "form-data": "^2.3.2",
+    "formidable": "^1.2.1",
+    "methods": "^1.1.2",
+    "mime": "^2.3.1",
+    "qs": "^6.5.2",
+    "readable-stream": "^3.0.4"
   },
   "devDependencies": {
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/preset-env": "^7.1.0",
+    "@commitlint/cli": "^7.1.2",
+    "@commitlint/config-conventional": "^7.1.2",
     "Base64": "^1.0.1",
-    "babel-core": "^6.26.3",
-    "babel-preset-es2015": "^6.24.1",
-    "babelify": "^8.0.0",
+    "accepts": "^1.3.5",
+    "babelify": "^10.0.0",
     "basic-auth-connect": "^1.0.0",
     "body-parser": "^1.18.2",
     "browserify": "^16.2.3",
+    "caniuse-lite": "^1.0.30000888",
+    "codecov": "^3.1.0",
+    "content-disposition": "^0.5.2",
+    "content-type": "^1.0.4",
+    "cookie": "^0.3.1",
     "cookie-parser": "^1.4.3",
+    "cookie-signature": "^1.1.0",
+    "cross-env": "^5.2.0",
+    "encoder": "^0.2.4",
+    "encodeurl": "^1.0.2",
+    "escape-html": "^1.0.3",
+    "eslint": "^5.6.1",
+    "eslint-config-prettier": "^3.1.0",
+    "eslint-config-xo-lass": "^1.0.3",
+    "eslint-plugin-compat": "^2.5.1",
+    "eslint-plugin-prettier": "^3.0.0",
+    "etag": "^1.8.1",
     "express": "^4.16.3",
     "express-session": "^1.15.6",
-    "marked": "^0.5.2",
+    "fixpack": "^2.3.1",
+    "fresh": "^0.5.2",
+    "husky": "^1.1.0",
+    "lint-staged": "^7.3.0",
+    "marked": "^0.5.0",
     "mocha": "^3.5.3",
-    "multer": "^1.4.1",
+    "multer": "^1.3.0",
+    "nyc": "^13.0.1",
+    "on-finished": "^2.3.0",
+    "parseurl": "^1.3.2",
+    "path-is-absolute": "^1.0.1",
+    "prettier": "^1.14.3",
+    "proxy-addr": "^2.0.4",
+    "range-parser": "^1.2.0",
+    "remark-cli": "^5.0.0",
+    "remark-preset-github": "^0.0.8",
+    "rimraf": "^2.6.2",
+    "safe-buffer": "^5.1.2",
+    "send": "^0.16.2",
     "should": "^13.2.0",
     "should-http": "^0.1.1",
+    "statuses": "^1.5.0",
+    "tag": "^0.4.1",
+    "tinyify": "^2.4.3",
+    "type-is": "^1.6.16",
+    "utils-merge": "^1.0.1",
+    "vary": "^1.1.2",
+    "xo": "^0.23.0",
     "zuul": "^3.12.0"
   },
-  "browser": {
-    "./lib/node/index.js": "./lib/client.js",
-    "./test/support/server.js": "./test/support/blank.js"
+  "engines": {
+    "node": ">=8.8.1"
   },
-  "component": {
-    "scripts": {
-      "superagent": "lib/client.js"
+  "files": [
+    "lib",
+    "dist"
+  ],
+  "homepage": "https://github.com/visionmedia/superagent",
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged && npm test",
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
     }
   },
-  "main": "./lib/node/index.js",
-  "engines": {
-    "node": ">= 6.0"
+  "jsdelivr": "dist/superagent.min.js",
+  "keywords": [
+    "ajax",
+    "api",
+    "cancel",
+    "client",
+    "http",
+    "https",
+    "promise",
+    "promises",
+    "request",
+    "requests",
+    "response",
+    "rest",
+    "retry",
+    "superagent",
+    "timeout",
+    "transform",
+    "xhr",
+    "xmlhttprequest"
+  ],
+  "license": "MIT",
+  "lint-staged": {
+    "linters": {
+      "*.js": [
+        "prettier --write",
+        "git add"
+      ],
+      "*.md,!CONTRIBUTING.md,!HISTORY.md,!docs": [
+        "remark . -qfo",
+        "git add"
+      ],
+      "package.json": [
+        "fixpack",
+        "git add"
+      ]
+    }
+  },
+  "main": "lib/node/request.js",
+  "nyc": {
+    "reporter": [
+      "lcov",
+      "html",
+      "text"
+    ]
+  },
+  "prettier": {
+    "singleQuote": true,
+    "bracketSpacing": true,
+    "trailingComma": "none"
+  },
+  "remarkConfig": {
+    "plugins": [
+      "preset-github"
+    ]
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/visionmedia/superagent"
+  },
+  "scripts": {
+    "browserify": "browserify lib/node/request.js -o dist/superagent.js -s superagent -d -t babelify",
+    "build": "npm run build:clean && npm run build:lib && npm run build:dist",
+    "build:clean": "rimraf lib dist",
+    "build:dist": "npm run browserify && npm run minify",
+    "build:lib": "babel src --out-dir lib",
+    "coverage": "nyc report --reporter=text-lcov > coverage.lcov && codecov",
+    "lint": "xo && remark . -qfo && eslint lib && eslint dist",
+    "make": "make test",
+    "make-test-http2": "make test-node-http2",
+    "minify": "cross-env NODE_ENV=production browserify lib/node/request.js -o dist/superagent.min.js -s superagent -t babelify -p tinyify",
+    "nyc": "cross-env NODE_ENV=test nyc make",
+    "test": "npm run build && npm run lint && npm run make",
+    "test-coverage": "npm run build && npm run lint && npm run nyc"
+  },
+  "unpkg": "dist/superagent.min.js",
+  "xo": {
+    "prettier": true,
+    "space": true,
+    "extends": [
+      "xo-lass"
+    ],
+    "envs": [
+      "node",
+      "browser"
+    ],
+    "globals": [
+      "ActiveXObject"
+    ],
+    "rules": {
+      "no-negated-condition": "off",
+      "no-multi-assign": "off",
+      "eqeqeq": "off",
+      "unicorn/catch-error-name": "off",
+      "guard-for-in": "off",
+      "no-eq-null": "off",
+      "no-useless-escape": "off",
+      "prefer-destructuring": "off",
+      "unicorn/explicit-length-check": "off",
+      "new-cap": "off",
+      "valid-jsdoc": "off",
+      "func-names": "off",
+      "complexity": "off",
+      "no-prototype-builtins": "off",
+      "no-unused-expressions": "off",
+      "no-sequences": "off",
+      "func-name-matching": "off",
+      "no-implicit-coercion": "off",
+      "promise/param-names": "off",
+      "default-case": "off",
+      "handle-callback-err": "off",
+      "import/order": "off",
+      "no-unused-vars": "off",
+      "block-scoped-var": "off",
+      "promise/prefer-await-to-then": "off",
+      "prefer-object-spread": "off"
+    },
+    "overrides": [
+      {
+        "files": "test/**/*.js",
+        "rules": {
+          "no-undef": "off",
+          "import/no-unassigned-import": "off",
+          "node/no-deprecated-api": "off",
+          "unicorn/filename-case": "off",
+          "no-var": "off",
+          "no-redeclare": "off",
+          "prefer-rest-params": "off",
+          "max-nested-callbacks": "off",
+          "promise/valid-params": "off",
+          "no-use-extend-native/no-use-extend-native": "off",
+          "no-extend-native": "off",
+          "no-void": "off"
+        }
+      }
+    ]
   }
 }
