@@ -9,7 +9,7 @@
  */
 
 function isObject(obj) {
-  return null !== obj && 'object' === typeof obj;
+  return obj !== null && typeof obj === 'object';
 }
 
 module.exports = isObject;
