@@ -1,16 +1,18 @@
 'use strict';
 
-module.exports = function parseJSON(res, fn){
+module.exports = function parseJSON(res, fn) {
   res.text = '';
   res.setEncoding('utf8');
   res.on('data', chunk => {
     res.text += chunk;
   });
   res.on('end', () => {
+    let body;
+    let err;
     try {
-      var body = res.text && JSON.parse(res.text);
+      body = res.text && JSON.parse(res.text);
     } catch (e) {
-      var err = e;
+      err = e;
       // issue #675: return the raw response if the response parsing fails
       err.rawResponse = res.text || null;
       // issue #876: return the http status code if the response parsing fails
