@@ -58,7 +58,7 @@ ResponseBase(Response.prototype);
  * Implements methods of a `ReadableStream`
  */
 
-Response.prototype.destroy = function(err){
+Response.prototype.destroy = function(err) {
   this.res.destroy(err);
 };
 
@@ -66,7 +66,7 @@ Response.prototype.destroy = function(err){
  * Pause.
  */
 
-Response.prototype.pause = function(){
+Response.prototype.pause = function() {
   this.res.pause();
 };
 
@@ -74,7 +74,7 @@ Response.prototype.pause = function(){
  * Resume.
  */
 
-Response.prototype.resume = function(){
+Response.prototype.resume = function() {
   this.res.resume();
 };
 
@@ -100,9 +100,8 @@ Response.prototype.toError = function() {
   return err;
 };
 
-
-Response.prototype.setStatusProperties = function(status){
-  console.warn("In superagent 2.x setStatusProperties is a private method");
+Response.prototype.setStatusProperties = function(status) {
+  console.warn('In superagent 2.x setStatusProperties is a private method');
   return this._setStatusProperties(status);
 };
 
@@ -118,6 +117,6 @@ Response.prototype.toJSON = function() {
     req: this.request.toJSON(),
     header: this.header,
     status: this.status,
-    text: this.text,
+    text: this.text
   };
 };
