@@ -1,12 +1,14 @@
-"use strict";
+'use strict';
 
-const request = require("../support/client");
-const express = require("../support/express");;
-const assert = require("assert");
+const request = require('../support/client');
+const express = require('../support/express');
+const assert = require('assert');
+
 const app = express();
-const url = require("url");
-const https = require("https");
-const fs = require("fs");
+const url = require('url');
+const https = require('https');
+const fs = require('fs');
+
 const ca = fs.readFileSync(`${__dirname}/fixtures/ca.cert.pem`);
 const key = fs.readFileSync(`${__dirname}/fixtures/key.pem`);
 const pfx = fs.readFileSync(`${__dirname}/fixtures/cert.pfx`);
@@ -28,36 +30,36 @@ openssl pkcs12 -export -in cert.pem -inkey key.pem -out passcert.pfx # password 
 
  */
 let http2;
-if(process.env.HTTP2_TEST){
+if (process.env.HTTP2_TEST) {
   http2 = require('http2');
 }
 let server;
 
-app.get("/", (req, res) => {
-  res.send("Safe and secure!");
+app.get('/', (req, res) => {
+  res.send('Safe and secure!');
 });
 
 // WARNING: this .listen() boilerplate is slightly different from most tests
 // due to HTTPS. Do not copy/paste without examination.
-const base = "https://localhost";
+const base = 'https://localhost';
 let testEndpoint;
 
-describe("https", () => {
-  describe("certificate authority", () => {
+describe('https', () => {
+  describe('certificate authority', () => {
     before(function listen(done) {
-      if(process.env.HTTP2_TEST){
+      if (process.env.HTTP2_TEST) {
         server = http2.createSecureServer(
           {
             key,
-            cert,
+            cert
           },
           app
         );
-      }else{
+      } else {
         server = https.createServer(
           {
             key,
-            cert,
+            cert
           },
           app
         );
@@ -72,15 +74,15 @@ describe("https", () => {
       if (server) server.close();
     });
 
-    describe("request", () => {
-      it("should give a good response", done => {
+    describe('request', () => {
+      it('should give a good response', done => {
         request
           .get(testEndpoint)
           .ca(ca)
           .end((err, res) => {
             assert.ifError(err);
             assert(res.ok);
-            assert.strictEqual("Safe and secure!", res.text);
+            assert.strictEqual('Safe and secure!', res.text);
             done();
           });
       });
@@ -108,17 +110,17 @@ describe("https", () => {
       });
     });
 
-    describe(".agent", () => {
-      it("should be able to make multiple requests without redefining the certificate", done => {
+    describe('.agent', () => {
+      it('should be able to make multiple requests without redefining the certificate', done => {
         const agent = request.agent({ ca });
         agent.get(testEndpoint).end((err, res) => {
           assert.ifError(err);
           assert(res.ok);
-          assert.strictEqual("Safe and secure!", res.text);
+          assert.strictEqual('Safe and secure!', res.text);
           agent.get(url.parse(testEndpoint)).end((err, res) => {
             assert.ifError(err);
             assert(res.ok);
-            assert.strictEqual("Safe and secure!", res.text);
+            assert.strictEqual('Safe and secure!', res.text);
             done();
           });
         });
@@ -126,27 +128,27 @@ describe("https", () => {
     });
   });
 
-  describe.skip("client certificates", () => {
+  describe.skip('client certificates', () => {
     before(function listen(done) {
-      if(process.env.HTTP2_TEST){
+      if (process.env.HTTP2_TEST) {
         server = http2.createSecureServer(
           {
             ca,
             key,
             cert,
             requestCert: true,
-            rejectUnauthorized: true,
+            rejectUnauthorized: true
           },
           app
         );
-      }else{
+      } else {
         server = https.createServer(
           {
             ca,
             key,
             cert,
             requestCert: true,
-            rejectUnauthorized: true,
+            rejectUnauthorized: true
           },
           app
         );
@@ -161,8 +163,8 @@ describe("https", () => {
       if (server) server.close();
     });
 
-    describe("request", () => {
-      it("should give a good response with client certificates and CA", done => {
+    describe('request', () => {
+      it('should give a good response with client certificates and CA', done => {
         request
           .get(testEndpoint)
           .ca(ca)
@@ -171,62 +173,62 @@ describe("https", () => {
           .end((err, res) => {
             assert.ifError(err);
             assert(res.ok);
-            assert.strictEqual("Safe and secure!", res.text);
+            assert.strictEqual('Safe and secure!', res.text);
             done();
           });
       });
-      it("should give a good response with client pfx", done => {
+      it('should give a good response with client pfx', done => {
         request
           .get(testEndpoint)
           .pfx(pfx)
           .end((err, res) => {
             assert.ifError(err);
             assert(res.ok);
-            assert.strictEqual("Safe and secure!", res.text);
+            assert.strictEqual('Safe and secure!', res.text);
             done();
           });
       });
-      it("should give a good response with client pfx with passphrase", done => {
+      it('should give a good response with client pfx with passphrase', done => {
         request
           .get(testEndpoint)
           .pfx({
             pfx: passpfx,
-            passphrase: "test",
+            passphrase: 'test'
           })
           .end((err, res) => {
             assert.ifError(err);
             assert(res.ok);
-            assert.strictEqual("Safe and secure!", res.text);
+            assert.strictEqual('Safe and secure!', res.text);
             done();
           });
       });
     });
 
-    describe(".agent", () => {
-      it("should be able to make multiple requests without redefining the certificates", done => {
+    describe('.agent', () => {
+      it('should be able to make multiple requests without redefining the certificates', done => {
         const agent = request.agent({ ca, key, cert });
         agent.get(testEndpoint).end((err, res) => {
           assert.ifError(err);
           assert(res.ok);
-          assert.strictEqual("Safe and secure!", res.text);
+          assert.strictEqual('Safe and secure!', res.text);
           agent.get(url.parse(testEndpoint)).end((err, res) => {
             assert.ifError(err);
             assert(res.ok);
-            assert.strictEqual("Safe and secure!", res.text);
+            assert.strictEqual('Safe and secure!', res.text);
             done();
           });
         });
       });
-      it("should be able to make multiple requests without redefining pfx", done => {
+      it('should be able to make multiple requests without redefining pfx', done => {
         const agent = request.agent({ pfx });
         agent.get(testEndpoint).end((err, res) => {
           assert.ifError(err);
           assert(res.ok);
-          assert.strictEqual("Safe and secure!", res.text);
+          assert.strictEqual('Safe and secure!', res.text);
           agent.get(url.parse(testEndpoint)).end((err, res) => {
             assert.ifError(err);
             assert(res.ok);
-            assert.strictEqual("Safe and secure!", res.text);
+            assert.strictEqual('Safe and secure!', res.text);
             done();
           });
         });
