@@ -1,8 +1,11 @@
-"use strict";
-const request = require("../support/client"),
-  express = require("../support/express"),
-  assert = require("assert"),
-  net = require("net");
+'use strict';
+const request = require('../support/client');
+
+const express = require('../support/express');
+
+const assert = require('assert');
+
+const net = require('net');
 
 function getFreePort(fn) {
   const server = net.createServer();
@@ -14,7 +17,7 @@ function getFreePort(fn) {
   });
 }
 
-describe("with network error", () => {
+describe('with network error', () => {
   before(function(done) {
     // connecting to a free port
     // will trigger a connection refused
@@ -24,9 +27,9 @@ describe("with network error", () => {
     });
   });
 
-  it("should error", function(done) {
+  it('should error', function(done) {
     request.get(`http://localhost:${this.port}/`).end((err, res) => {
-      assert(err, "expected an error");
+      assert(err, 'expected an error');
       done();
     });
   });
