@@ -1,16 +1,18 @@
-"use strict";
-const request = require("../support/client"),
-  express = require("../support/express"),
-  app = express();
+'use strict';
+const request = require('../support/client');
+
+const express = require('../support/express');
+
+const app = express();
 let http = require('http');
-let assert = require('assert');
 
-describe("request.get().set()", () => {
-  if (process.env.HTTP2_TEST) {
-    return; // request object doesn't look the same
-  }
-  let server;
+if (process.env.HTTP2_TEST) {
+  http = require('http2');
+}
 
+let base = 'http://localhost';
+let server;
+describe('request.get().set()', () => {
   after(function exitServer() {
     if (typeof server.close === 'function') {
       server.close();
@@ -19,27 +21,22 @@ describe("request.get().set()", () => {
     }
   });
 
-  it("should set host header after get()", done => {
-    app.get("/", (req, res) => {
-      assert.equal(req.hostname, 'example.com');
+  it('should set host header after get()', done => {
+    app.get('/', (req, res) => {
+      req.hostname.should.equal('example.com');
       res.end();
     });
 
     server = http.createServer(app);
     server.listen(0, function listening() {
+      base += `:${server.address().port}`;
 
       request
-        .get(`http://localhost:${server.address().port}`)
+        .get(base)
         .set('host', 'example.com')
-        .then(() => {
-          return request
-            .get(`http://example.com:${server.address().port}`)
-            .connect({
-              "example.com": "localhost",
-              "*": "fail",
-            });
-        })
-        .then(() => done(), done);
+        .end(() => {
+          done();
+        });
     });
   });
 });
