@@ -1,17 +1,18 @@
-"use strict";
-const request = require("../support/client");
-const express = require("../support/express");
-const assert = require("assert");
+'use strict';
+const request = require('../support/client');
+const express = require('../support/express');
+const assert = require('assert');
+
 const app = express();
-let http = require("http");
-let https = require("https");
-const os = require("os");
-const fs = require("fs");
+let http = require('http');
+let https = require('https');
+const os = require('os');
+const fs = require('fs');
+
 const key = fs.readFileSync(`${__dirname}/fixtures/key.pem`);
 const cert = fs.readFileSync(`${__dirname}/fixtures/cert.pem`);
-const cacert = fs.readFileSync(`${__dirname}/fixtures/ca.cert.pem`);
-const httpSockPath = [os.tmpdir(), "superagent-http.sock"].join("/");
-const httpsSockPath = [os.tmpdir(), "superagent-https.sock"].join("/");
+const httpSockPath = [os.tmpdir(), 'superagent-http.sock'].join('/');
+const httpsSockPath = [os.tmpdir(), 'superagent-https.sock'].join('/');
 let httpServer;
 let httpsServer;
 
@@ -19,16 +20,16 @@ if (process.env.HTTP2_TEST) {
   http = https = require('http2');
 }
 
-app.get("/", (req, res) => {
-  res.send("root ok!");
+app.get('/', (req, res) => {
+  res.send('root ok!');
 });
 
-app.get("/request/path", (req, res) => {
-  res.send("request path ok!");
+app.get('/request/path', (req, res) => {
+  res.send('request path ok!');
 });
 
-describe("[unix-sockets] http", () => {
-  if (process.platform === "win32") {
+describe('[unix-sockets] http', () => {
+  if (process.platform === 'win32') {
     return;
   }
 
@@ -41,21 +42,21 @@ describe("[unix-sockets] http", () => {
     httpServer.listen(httpSockPath, done);
   });
 
-  const base = `http+unix://${httpSockPath.replace(/\//g, "%2F")}`;
+  const base = `http+unix://${httpSockPath.replace(/\//g, '%2F')}`;
 
-  describe("request", () => {
-    it("path: / (root)", done => {
+  describe('request', () => {
+    it('path: / (root)', done => {
       request.get(`${base}/`).end((err, res) => {
         assert(res.ok);
-        assert.strictEqual("root ok!", res.text);
+        assert.strictEqual('root ok!', res.text);
         done();
       });
     });
 
-    it("path: /request/path", done => {
+    it('path: /request/path', done => {
       request.get(`${base}/request/path`).end((err, res) => {
         assert(res.ok);
-        assert.strictEqual("request path ok!", res.text);
+        assert.strictEqual('request path ok!', res.text);
         done();
       });
     });
@@ -64,13 +65,12 @@ describe("[unix-sockets] http", () => {
   after(() => {
     if (typeof httpServer.close === 'function') {
       httpServer.close();
-    } else
-      httpServer.destroy();
+    } else httpServer.destroy();
   });
 });
 
-describe("[unix-sockets] https", () => {
-  if (process.platform === "win32") {
+describe('[unix-sockets] https', () => {
+  if (process.platform === 'win32') {
     return;
   }
 
@@ -87,29 +87,27 @@ describe("[unix-sockets] https", () => {
     httpsServer.listen(httpsSockPath, done);
   });
 
-  const base = `https+unix://${httpsSockPath.replace(/\//g, "%2F")}`;
+  const base = `https+unix://${httpsSockPath.replace(/\//g, '%2F')}`;
 
-  describe("request", () => {
-    it("path: / (root)", done => {
+  describe('request', () => {
+    it('path: / (root)', done => {
       request
         .get(`${base}/`)
-        .ca(cacert)
+        .ca(cert)
         .end((err, res) => {
-          assert.ifError(err);
           assert(res.ok);
-          assert.strictEqual("root ok!", res.text);
+          assert.strictEqual('root ok!', res.text);
           done();
         });
     });
 
-    it("path: /request/path", done => {
+    it('path: /request/path', done => {
       request
         .get(`${base}/request/path`)
-        .ca(cacert)
+        .ca(cert)
         .end((err, res) => {
-          assert.ifError(err);
           assert(res.ok);
-          assert.strictEqual("request path ok!", res.text);
+          assert.strictEqual('request path ok!', res.text);
           done();
         });
     });
