@@ -6,8 +6,7 @@ let uri = 'http://localhost:5000';
 if (typeof window !== 'undefined') {
   NODE = false;
   uri = `//${window.location.host}`;
-}
-else {
+} else {
   process.env.ZUUL_PORT = 5000;
   require('./server');
 }
