@@ -1,20 +1,21 @@
 const setup = require('./support/setup');
+
 const uri = setup.uri;
 
 const assert = require('assert');
 const request = require('./support/client');
 
-describe('request', function(){
+describe('request', function() {
   this.timeout(20000);
   describe('use', () => {
     it('should use plugin success', done => {
       const now = `${Date.now()}`;
-      function uuid(req){
+      function uuid(req) {
         req.set('X-UUID', now);
         return req;
       }
-      function prefix(req){
-        req.url = uri + req.url
+      function prefix(req) {
+        req.url = uri + req.url;
         return req;
       }
       request
@@ -25,10 +26,10 @@ describe('request', function(){
           assert.strictEqual(res.statusCode, 200);
           assert.equal(res.get('X-UUID'), now);
           done();
-        })
+        });
     });
   });
-})
+});
 
 describe('subclass', () => {
   let OriginalRequest;
@@ -47,7 +48,8 @@ describe('subclass', () => {
   it('should use patched subclass', () => {
     assert(OriginalRequest);
 
-    let constructorCalled, sendCalled;
+    let constructorCalled;
+    let sendCalled;
     function NewRequest(...args) {
       constructorCalled = true;
       OriginalRequest.apply(this, args);
@@ -80,4 +82,4 @@ describe('subclass', () => {
     assert(req instanceof NewRequest);
     assert(req instanceof OriginalRequest);
   });
-})
+});
